Update to VRS 1.2, remove deprecated models, rename model to models
