+++ tests/conftest.py
@@ -1,10 +1,10 @@
 """Module for pytest config tools."""
 import pytest
-from ga4gh.vrsatile.pydantic.vrs_model import SequenceInterval, \
+from ga4gh.vrsatile.pydantic.vrs_models import SequenceInterval, \
     CytobandInterval, SequenceLocation, DerivedSequenceExpression, Number, \
     IndefiniteRange, DefiniteRange, Allele, LiteralSequenceExpression, Gene, \
     ChromosomeLocation
-from ga4gh.vrsatile.pydantic.vrsatile_model import Extension, Expression, \
+from ga4gh.vrsatile.pydantic.vrsatile_models import Extension, Expression, \
     SequenceDescriptor, LocationDescriptor, GeneDescriptor, VCFRecord
 
 
@@ -49,7 +49,7 @@ def chromosome_location(cytoband_interval):
     return ChromosomeLocation(
         chr="19",
         interval=cytoband_interval,
-        species_id="taxonomy:9606"
+        species="taxonomy:9606"
     )
